requirements.txt
asgiref==3.6.0
cfgv==3.3.1
distlib==0.3.6
dj-database-url==1.2.0
Django==4.1.4
django-phonenumber-field==7.0.1
filelock==3.9.0
identify==2.5.11
nodeenv==1.7.0
phonenumberslite==8.13.3
platformdirs==2.6.2
pre-commit==2.21.0
psycopg2-binary==2.9.5
python-dotenv==0.21.0
PyYAML==6.0
sqlparse==0.4.3
virtualenv==20.17.1
